Update api's to load paged data.
Plants data is loaded in pages using paging library. Flow
MutableStateFlow is used to collect the paged data in
repository and viewmodel.

+++ app/src/main/java/com/jassemdev/android/plants/data/PlantsApi.kt
@@ -36,7 +36,7 @@ interface PlantsApi {
     suspend fun getPlants(@Query("token") token: String): PlantsApiModel
 
     @GET("plants")
-    suspend fun getPlantsByPage(@Query("page") pageNumber: Long,
+    suspend fun getPlantsByPage(@Query("page") pageNumber: Int,
                                 @Query("token") token: String): PlantsApiModel
 
     @GET("plants/{id}")
